% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{addEsriImageMapLayer}
\alias{addEsriImageMapLayer}
\title{woprVision: Leaflet map helper}
\usage{
addEsriImageMapLayer(map, url, layerId = NULL, group = NULL, options = NULL)
}
\arguments{
\item{map}{A map widget object created from leaflet()}

\item{url}{URL of ImageServer tiles server}

\item{layerId}{tile layer Id}

\item{group}{The name of the group the newly created layers should belong to}

\item{opacity}{layer opacity (int between 0 and 1)}
}
\value{
A Leaflet map.
}
\description{
Display ESRI ImageServer tiles for the Leaflet in woprVision
}
