% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellids.R
\name{cellids}
\alias{cellids}
\title{Get cell ids from point or polygon}
\usage{
cellids(feature, mastergrid)
}
\arguments{
\item{feature}{An object of class sf with a polygon or point}

\item{mastergrid}{A mastergrid raster downloaded from WOPR}
}
\value{
A numeric vector containing cell IDs
}
\description{
Returns cell ids for a polygon or point location that can be used to query a WOPR SQL database.
}
