% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLocal.R
\name{checkLocal}
\alias{checkLocal}
\title{Check for local WOPR files}
\usage{
checkLocal(dir, info = wopr:::woprVision_global$version_info)
}
\arguments{
\item{dir}{Directory where WOPR files are stored}

\item{info}{Data "version_info" object to update with local files (see wopr:::woprVision_global$version_info)}
}
\description{
Check for local files required to run woprVision in local mode (without internet connection)
}
