% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTask.R
\name{checkTask}
\alias{checkTask}
\title{Check WOPR task}
\usage{
checkTask(taskid, production = F)
}
\arguments{
\item{taskid}{A character vector of task IDs}

\item{production}{Logical indicating whether to use the test server or production server}
}
\value{
A list with a result for each task id which will be a vector of samples from posterior distribution of the population total if that task has completed.
}
\description{
Retrieve results from a task id or task ids that previously timed out
}
