% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyWP.R
\name{copyWP}
\alias{copyWP}
\title{Copy WorldPop folder to your local working directory.}
\usage{
copyWP(srcdir, outdir, overwrite = F, OS.type = "windows")
}
\arguments{
\item{srcdir}{Directory of source files}

\item{outdir}{Directory to write output files to}

\item{overwrite}{Logical indicating to overwrite local files if they already exist}

\item{OS.type}{Type of operating system (see \code{.Platform$OS.type})}
}
\value{
Writes input files to disk
}
\description{
Copy files from WorldPop server to your local working directory. 
This allows everyone to work locally with identical inputs.
}
