% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadData}
\alias{downloadData}
\title{Download data from WOPR}
\usage{
downloadData(dat, outdir, maxsize = 50, dialogue = T)
}
\arguments{
\item{dat}{Data sets to download provided as row or rows from the data catalogue (see ?getCatalogue)}

\item{outdir}{Directory where downloads should be saved}

\item{maxsize}{Maximum file size (MB) allowed to download without notification}

\item{dialogue}{Logical. If TRUE a dialogue box will ask if you want to download any file that exceeds maxsize}
}
\value{
Files are downloaded directly to local disk
}
\description{
Download data from WOPR
}
