% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeSexTable.R
\name{getAgeSexNames}
\alias{getAgeSexNames}
\title{Extract agesex group labels from an agesex table}
\usage{
getAgeSexNames(agesex_table)
}
\arguments{
\item{agesex_table}{Table retrieved with getAgeSexTable}
}
\value{

}
\description{
Extract agesex group labels from an agesex table
}
