% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeSexTable.R
\name{getAgeSexTable}
\alias{getAgeSexTable}
\title{Get country-specific age sex proportions either through WOPR API or locally}
\usage{
getAgeSexTable(country, version = NA, locator = "https://api.worldpop.org/v1")
}
\arguments{
\item{country}{The ISO3 country code}

\item{version}{Version number of the population estimate}

\item{locator}{Server url or local path to age sex table (optional)}
}
\value{
A table with age and sex proportions for the selected country and version
}
\description{
Get country-and-version-specific age and sex proportions through querying WOPR API or uploading local age sex table
}
