% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatalogue.R
\name{getCatalogue}
\alias{getCatalogue}
\title{Get the WorldPop GridFree data catalogue}
\usage{
getCatalogue()
}
\value{
A data frame with a row for each item in the data catalogue.
}
\description{
Get the WorldPop GridFree data catalogue
}
