% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPop.R
\name{getPop}
\alias{getPop}
\title{Get population estimate from GRID3 server via API request}
\usage{
getPop(geojson, iso3, ver)
}
\arguments{
\item{geojson}{A geoJSON to represent the polygon where a population estimate is needed}

\item{iso3}{The ISO3 country code}

\item{ver}{The version of the population estimate}

\item{alpha}{The type 1 error rate for the confidence intervals}

\item{tails}{The number of tails for the confidence intervals}
}
\value{
A list with the mean, median, lower and upper confidence interval for the population total
}
\description{
Get population estimate from GRID3 server via API request
}
