% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopSql.R
\name{getPopSql}
\alias{getPopSql}
\title{Query local SQL database}
\usage{
getPopSql(cells, db)
}
\arguments{
\item{cells}{A numeric vector of cell ids (see ?cellids)}

\item{db}{An SQLite database connection}
}
\value{
Estimated population total summed across all cell ids. Result is returned as a numeric vector containing samples from the predicted posterior distribution.
}
\description{
Returns population estimates from a WOPR SQL database based on cell ids provided.
}
