% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken_ESRI.R
\name{getToken_ESRI}
\alias{getToken_ESRI}
\title{Generate token from username/password to access password-protected population layer from ESRI ImageService}
\usage{
getToken_ESRI(u, p)
}
\arguments{
\item{u}{chr Username}

\item{p}{chr Password}
}
\value{
chr Generated token
}
\description{
Generate token from username/password to access password-protected population layer from ESRI ImageService
}
