% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_sf.R
\name{leaf_sf}
\alias{leaf_sf}
\title{Convert a leaflet drawn polygon to shapefile}
\usage{
leaf_sf(geometry, pointpoly = "Selected Point")
}
\arguments{
\item{geometry}{leaflet polygon}

\item{pointpoly}{location selection mode ('Selected Point' or 'Custom Area')}
}
\description{
Convert leaflet polygon to sf_POLYGONS
}
