% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{woprVision: Leaflet map}
\usage{
map(country, version, local_tiles = F, southern = F)
}
\arguments{
\item{country}{ISO3 code for country to map}

\item{version}{Version of data set to map}

\item{local_tiles}{Logical indicating to use locally-stored tiles to display the population raster}

\item{southern}{Logical indicating if the country is in the southern hemisphere (to use "{-y}" format for Leaflet tiles)}
}
\value{
A Leaflet map.
}
\description{
Leaflet map for woprVision.
}
