% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapProxyFile.R
\name{mapProxyFile}
\alias{mapProxyFile}
\title{Map proxy to make changes when a GeoJSON is uploaded}
\usage{
mapProxyFile(features, map_zoom)
}
\arguments{
\item{features}{sf features}

\item{map_zoom}{current leaflet map zoom level}
}
\description{
Update map for polygon mode
}
