% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapProxyMarker.R
\name{mapProxyMarker}
\alias{mapProxyMarker}
\title{Leaflet map proxy to update on map click}
\usage{
mapProxyMarker(pointpoly, map_click, map_zoom)
}
\value{
A n updated leaflet map.
}
\description{
Leaflet map proxy that zooms to map click and drops a pin.
}
