% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapProxyPoly.R
\name{mapProxyPoly}
\alias{mapProxyPoly}
\title{Map proxy to make changes when polygon mode is selected}
\usage{
mapProxyPoly(pointpoly)
}
\arguments{
\item{pointpoly}{Tool selection ('Selected Point' or 'Custom Area')}
}
\description{
Update map for polygon mode
}
