% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFilename.R
\name{parseFilename}
\alias{parseFilename}
\title{Parse WOPR file names}
\usage{
parseFilename(filenames)
}
\arguments{
\item{filenames}{character vector. WOPR file name.}
}
\value{
data.frame. Table of results.
}
\description{
Parse WOPR file names to return the country, category, version, file type, optional info, and the file extension.
}
