% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPanel.R
\name{plotPanel}
\alias{plotPanel}
\title{woprVision: create a panel to hold plots}
\usage{
plotPanel(
  N,
  agesexSelect,
  agesexTable = NULL,
  confidence = 95,
  tails = "Interval",
  popthresh = 100
)
}
\description{
Produces a panel to hold the population density plot and the population pyramid
}
