% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.R
\name{plotPop}
\alias{plotPop}
\title{Plot population estimate}
\usage{
plotPop(
  N,
  confidence = 95,
  tails = "Interval",
  popthresh = 100,
  popmax = NA,
  dict = dict_en
)
}
\arguments{
\item{N}{Numerical vector. Population estimate from WOPR.}

\item{confidence}{The confidence level for the confidence intervals (i.e. percentage from 0 to 100)}

\item{tails}{The confidence type: 'Interval', 'Lower Threshold', or 'Upper Threshold'}

\item{popthresh}{The function will return the probability that the population size exceeds _popthresh_}

\item{popmax}{numeric. Maximum population size for a pixel in the country}

\item{dict}{Dictionnary for text translation}
}
\description{
Plot population estimate
}
