% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPyramid.R
\name{plotPyramid}
\alias{plotPyramid}
\title{Create an age-sex population pyramid}
\usage{
plotPyramid(agesexSelect, agesexTable)
}
\arguments{
\item{agesexSelect}{a character vector of selected Age and Sex groups}

\item{agesexTable}{The age sex table}
}
\description{
Produces an age sex pyramid plot with the selected age sex classes highlighted.
}
