% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultTable.R
\name{resultTable}
\alias{resultTable}
\title{woprVision: Create a table to store results}
\usage{
resultTable(inp, rval)
}
\arguments{
\item{inp}{input object from ui}

\item{rval}{reactiveValues object from server}
}
\description{
Save results from selected locations
}
