% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveResults.R
\name{retrieveResults}
\alias{retrieveResults}
\title{Retrieve results from WOPR}
\usage{
retrieveResults(
  tasks,
  url,
  confidence = 0.95,
  tails = 2,
  popthresh = NA,
  summarize = T,
  timeout = 30 * 60,
  verbose = T
)
}
\arguments{
\item{tasks}{A data frame with information about tasks (see ?submitTasks)}

\item{url}{The url of the WorldPop queue to check for results}

\item{confidence}{The confidence level for the confidence intervals (e.g. 0.95 = 95 percent confidence intervals)}

\item{tails}{The number of tails for the confidence intervals}

\item{popthresh}{Threshold population size to calculate the probability that population exceeds}

\item{summarize}{Logical indicating to summarize results or return all posterior samples}

\item{timeout}{Seconds until function times out}

\item{verbose}{Logical indicating to print progress updates}
}
\value{
A data frame with outputs
}
\description{
Retrieve results from WOPR
}
