% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPop.R
\name{summaryPop}
\alias{summaryPop}
\title{Summary statistics}
\usage{
summaryPop(
  N,
  confidence = 0.95,
  tails = 2,
  belowthresh = NA,
  abovethresh = NA,
  round_result = F
)
}
\arguments{
\item{N}{numeric vector. Vector of posterior samples for the population total}

\item{confidence}{numeric. The confidence level for the confidence intervals (e.g. 0.95 = 95 percent confidence intervals)}

\item{tails}{integer. The number of tails for the confidence intervals}

\item{belowthresh}{numeric. The function will return the probability that the population size is less than _belowthresh_}

\item{abovethresh}{numeric. The function will return the probability that the population size exceeds _abovethresh_}

\item{round_result}{logical. If TRUE the results will be rounded to a pre-determined number of digits for each column.}
}
\value{
A data.frame with columns containing the mean, median, lower and upper confidence intervals for the estimated population total. 
The 'abovethresh' column reports the probability that the population is greater than _abovethresh_. 
The 'belowthresh' column reports the probability that the population is less than _abovethresh_.
Note: One minus 'abovethresh' is the probability that the population is equal to or less than _abovethresh_.
}
\description{
Summarize predicted posterior probability distribution for population estimates
}
