% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wopr_points-data.R
\docType{data}
\name{wopr_points}
\alias{wopr_points}
\title{Example points for WOPR}
\format{
An object of type _sf_: sfc_POINTS and sfc_MULTIPOINTS
}
\usage{
data(wopr_points)
}
\description{
Points of class _sf_ to use as example data for _wopr_ package. These points are within Nigeria (country='NGA')
}
\keyword{datasets}
