% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wopr_polys-data.R
\docType{data}
\name{wopr_polys}
\alias{wopr_polys}
\title{Example polygons for WOPR}
\format{An object of type _sf_: sfc_POLYGONS and sfc_MULTIPOLYGONS}
\usage{
data(wopr_polys)
}
\description{
Polygons of class _sf_ to use as example data for _wopr_ package. These polygons are within Nigeria (country='NGA')
}
\keyword{datasets}
