% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woprize.R
\name{woprize}
\alias{woprize}
\title{WOPRize a shapefile}
\usage{
woprize(
  features,
  country,
  version = NA,
  confidence = 0.95,
  tails = 2,
  abovethresh = NA,
  belowthresh = NA,
  summarize = T,
  agesex_select = c(paste0("m", c(0, 1, seq(5, 80, 5))), paste0("f", c(0, 1, seq(5, 80,
    5)))),
  key = "key.txt",
  save_messages = F,
  url = "https://api.worldpop.org/v1"
)
}
\arguments{
\item{features}{sf. An object of class sf with points or polygons to calculate population totals.}

\item{country}{character. ISO-3 code for the country requested.}

\item{version}{character or numeric. Version number of population estimates.}

\item{confidence}{numeric. The confidence level for the confidence intervals (e.g. 0.95 = 95 percent confidence intervals).}

\item{tails}{numeric. The number of tails for the confidence intervals.}

\item{abovethresh}{numeric. The function will return the probability that the population size exceeds _abovethresh_.}

\item{belowthresh}{numeric. The function will return the probability that the population size is less than _belowthresh_.}

\item{summarize}{logical. If TRUE, results will be summarized. If FALSE, all posterior samples will be returned.}

\item{agesex_select}{charcter vector. Age-sex groups (e.g. "m0", "m1", "m5").}

\item{key}{character. API key.}

\item{save_messages}{logical. If TRUE, messages from WOPR will be saved including task ids that can be used to retrieve results later using checkTask().}

\item{url}{character. Server url (optional).}
}
\value{
sf. An sf spatial data object with summaries of posterior distributions for estimates of total population within each polygon.
}
\description{
Query WOPR to get population totals and confidence intervals for sf polygon or point features.
}
