% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpEndpoint.R
\name{wpEndpoint}
\alias{wpEndpoint}
\title{Identify url for appropriate WorldPop API endpoint}
\usage{
wpEndpoint(geometry_class, agesex = T, production = F)
}
\arguments{
\item{geometry_class}{Class of geometry object being submitted}

\item{agesex}{Logical indicating whether population totals are needed for selected age-sex groups}

\item{production}{Logical indicating if production or development servers should be used}
}
\value{
A list with urls for the endpoint and the queue
}
\description{
Identify url for appropriate WorldPop API endpoint
}
