% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCatalogue.R
\name{writeCatalogue}
\alias{writeCatalogue}
\title{Writes the data catalogue to file based on the files currently on the local disk}
\usage{
writeCatalogue(outdir)
}
\arguments{
\item{outdir}{Output folder where data from the WorldPop GridFree catalogue are stored}
}
\value{
Writes an updated data catalogue to disk as a .csv file.
}
\description{
Writes the data catalogue to file based on the files currently on the local disk
}
