% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesexLookup.R
\name{agesexLookup}
\alias{agesexLookup}
\title{Format shiny age-group selection for WOPR}
\usage{
agesexLookup(male, female, male_select, female_select)
}
\arguments{
\item{ui}{user input from ui}
}
\description{
Format shiny age-group selection for WOPR
}
