% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTask.R
\name{checkTask}
\alias{checkTask}
\title{Retrieve results from a task that timed out}
\usage{
checkTask(taskid, test = F)
}
\arguments{
\item{taskid}{The task ID returned from requestPop()}

\item{test}{Logical indicating whether to use the test server or production server}
}
\value{
A vector of samples from posterior distribution of the population total. If the task times out the function will return the task ID.
}
\description{
Retrieve results from a task that timed out
}
