% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyWP.R
\name{copyWP}
\alias{copyWP}
\title{WorldPop Mirror}
\usage{
copyWP(srcdir, outdir, OS.type = "windows")
}
\arguments{
\item{srcdir}{Directory of source files}

\item{outdir}{Directory to write output files to}

\item{OS.type}{Type of operating system (see \code{.Platform$OS.type})}
}
\value{
Writes input files to disk
}
\description{
Mirror files from WorldPop drive to your local working directory.
}
