% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadData}
\alias{downloadData}
\title{Download data from WOPR}
\usage{
downloadData(dat, wopr_dir = "wopr", maxsize = 100)
}
\arguments{
\item{dat}{Data sets to download provided as a data.frame with at least one row from the wOPR data catalogue (see ?getCatalogue)}

\item{wopr_dir}{Directory where downloads should be saved}

\item{maxsize}{Maximum file size (MB) allowed to download without notification}
}
\value{
Files are downloaded directly to local disk
}
\description{
Download data from WOPR
}
