% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{endpoint}
\alias{endpoint}
\title{Identify url for appropriate WOPR endpoint}
\usage{
endpoint(features, agesex = T, production = F)
}
\arguments{
\item{agesex}{Logical indicating whether population totals are needed for selected age and sex groups}

\item{production}{Logical indicating if production or development servers should be used}

\item{geometry_class}{Class of geometry object being submitted}
}
\value{
A list with urls for the endpoint and the queue
}
\description{
Identify url for appropriate WOPR endpoint
}
