% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatalogue.R
\name{getCatalogue}
\alias{getCatalogue}
\title{WOPR data catalogue}
\usage{
getCatalogue(spatialQuery = F)
}
\value{
A data frame with a row for each item in the data catalogue.
}
\description{
Get a complete listing of data in the WOPR data catalogue.
}
