% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatalogue.R
\name{getCatalogue}
\alias{getCatalogue}
\title{WOPR data catalogue}
\usage{
getCatalogue(spatial_query = F)
}
\arguments{
\item{spatial_query}{Logical. If TRUE will return the data sets that support a spatial query (i.e. contain an SQL database). If FALSE will return the entire data catalogue.}
}
\value{
A data frame with a row for each item in the data catalogue.
}
\description{
Get a complete listing of data in the WOPR data catalogue.
}
