% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPop.R
\name{getPop}
\alias{getPop}
\title{Get population estimate for a single point or polygon}
\usage{
getPop(
  feature,
  country,
  ver = NA,
  agesex = c(paste0("m", c(0, 1, seq(5, 80, 5))), paste0("f", c(0, 1, seq(5, 80, 5)))),
  timeout = 5 * 60,
  key = "key.txt",
  verbose = T,
  getAgesexId = F,
  url = NA
)
}
\arguments{
\item{feature}{An object of class sf with a point or polygon to calculate population total. If more than one feature are included, only the first feature will be processed.}

\item{country}{The ISO3 country code}

\item{ver}{Version number of the population estimate}

\item{agesex}{Character vector of age-sex groups}

\item{timeout}{Seconds until timeout}

\item{key}{Key to increase daily quota for REST API requests}

\item{verbose}{Logical. Toggles on status updates while processing}

\item{url}{Server url (optional)}

\item{agesexid}{Logical. Toggle return of numeric age sex id (for woprVision)}
}
\value{
A vector of samples from posterior distribution of the population total. If the task times out the function will return the task ID.
}
\description{
Query WOPR using an sf point or polygon to get population estimates as a vector of samples from the Bayesian posterior prediction for that location.
}
