% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPop.R
\name{getPop}
\alias{getPop}
\title{Get population estimate from GRID3 server via API request}
\usage{
getPop(geojson, iso3, ver)
}
\arguments{
\item{geojson}{A geoJSON to represent the polygon where a population estimate is needed}

\item{iso3}{The ISO3 country code}

\item{ver}{The version of the population estimate}
}
\value{
A vector of samples from posterior distributio of the population total
}
\description{
Get population estimate from GRID3 server via API request
}
