% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPop.R
\name{getPop}
\alias{getPop}
\title{Get population estimate for a single point or polygon}
\usage{
getPop(
  feature,
  country,
  ver = NA,
  agesex = c("m0", "m1", "m5", "m10", "m15", "m20", "m25", "m30", "m35", "m40", "m45",
    "m50", "m55", "m60", "m65", "m70", "m75", "m80", "f0", "f1", "f5", "f10", "f15",
    "f20", "f25", "f30", "f35", "f40", "f45", "f50", "f55", "f60", "f65", "f70", "f75",
    "f80"),
  timeout = 60,
  key = "key.txt",
  verbose = T
)
}
\arguments{
\item{feature}{An object of class sf with a point or polygon to calculate population total. If more than one feature are included, only the first feature will be processed.}

\item{country}{The ISO3 country code}

\item{ver}{Version number of the population estimate}

\item{agesex}{Character vector of age-sex groups}

\item{timeout}{Seconds until timeout}

\item{key}{Key to increase daily quota for REST API requests}
}
\value{
A vector of samples from posterior distribution of the population total. If the task times out the function will return the task ID.
}
\description{
Query WOPR using an sf point or polygon to get population estimates as a vector of samples from the Bayesian posterior prediction for that location.
}
