% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopSql.R
\name{getPopSql}
\alias{getPopSql}
\title{Query local SQL database}
\usage{
getPopSql(
  cells,
  db,
  agesexTable,
  getAgesexId = F,
  verbose = T,
  block_size = 999,
  agesexSelect = c(paste0("m", c(0, 1, seq(5, 80, 5))), paste0("f", c(0, 1, seq(5, 80,
    5))))
)
}
\arguments{
\item{cells}{A numeric vector of cell ids (see ?cellids)}

\item{db}{An SQLite database connection}

\item{getAgesexId}{Logical indicating to return the agesex region ID.}

\item{verbose}{Logical indicating to print status updates while processing}

\item{agesexSelect}{Character vector of age-sex groups (e.g. c(m0,m1,m5,m10,m15,f0,f1) includes boys 0-19 and girls under 5)}
}
\value{
Estimated population total summed across all cell ids. Result is returned as a numeric vector containing samples from the predicted posterior distribution.
}
\description{
Returns population estimates from a WOPR SQL database based on cell ids provided.
}
