% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopSql.R
\name{getPopSql}
\alias{getPopSql}
\title{Query local SQL database}
\usage{
getPopSql(
  cells,
  db,
  agesex_table = NA,
  get_agesexid = F,
  verbose = T,
  max_area = 1000,
  timeout = 60,
  agesex_select = c(paste0("f", c(0, 1, seq(5, 80, 5))), paste0("m", c(0, 1, seq(5, 80,
    5))))
)
}
\arguments{
\item{cells}{A numeric vector of cell ids (see ?cellids)}

\item{db}{An SQLite database connection}

\item{agesex_table}{The age sex table}

\item{get_agesexid}{Logical indicating to return the agesex region ID.}

\item{verbose}{Logical indicating to print status updates while processing}

\item{max_area}{Maximum area (sq km) of polygons allowed to query the SQL database}

\item{timeout}{Timeout period (seconds)}

\item{agesex_select}{Character vector of age-sex groups (e.g. c(m0,m1,m5,m10,m15,f0,f1) includes boys 0-19 and girls under 5)}
}
\value{
Estimated population total summed across all cell ids. Result is returned as a numeric vector containing samples from the predicted posterior distribution.
}
\description{
Returns population estimates from a WOPR SQL database based on cell ids provided.
}
