% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{woprVision: Leaflet map}
\usage{
map(
  country,
  version,
  bins = wopr:::woprVision_global$bins,
  pal = wopr:::woprVision_global$pal,
  dict = dict_EN,
  token = NULL
)
}
\arguments{
\item{country}{ISO3 code for country to map}

\item{version}{Version of data set to map}

\item{bins}{Color bins for legend}

\item{pal}{Function for legend color palette (see ?leaflet::colorBin)}

\item{dict}{Dictionary for text translation}

\item{token}{Token generated for access to password-protected datasets}
}
\value{
A Leaflet map.
}
\description{
Leaflet map for woprVision.
}
