% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{woprVision: Leaflet map}
\usage{
map(country, version, dict = dict_EN, token = NULL, esri_server = F)
}
\arguments{
\item{country}{ISO3 code for country to map}

\item{version}{Version of data set to map}

\item{dict}{Dictionary for text translation}

\item{token}{Token generated for access to password-protected datasets}

\item{esri_server}{T/F to ping population tiles from esri server (more zoom capacity) or from wolrdpop server}
}
\value{
A Leaflet map.
}
\description{
Leaflet map for woprVision.
}
