% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{woprVision: Leaflet map}
\usage{
map(
  country,
  version,
  local_tiles = F,
  southern = F,
  bins = wopr:::woprVision_global$bins,
  pal = wopr:::woprVision_global$pal
)
}
\arguments{
\item{country}{ISO3 code for country to map}

\item{version}{Version of data set to map}

\item{local_tiles}{Logical indicating to use locally-stored tiles to display the population raster}

\item{southern}{Logical indicating if the country is in the southern hemisphere (to use "{-y}" format for Leaflet tiles)}

\item{bins}{Color bins for legend}

\item{pal}{Function for legend color palette (see ?leaflet::colorBin)}
}
\value{
A Leaflet map.
}
\description{
Leaflet map for woprVision.
}
