% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapProxyFile.R
\name{mapProxyFile}
\alias{mapProxyFile}
\title{Map proxy to make changes when a GeoJSON is uploaded}
\usage{
mapProxyFile(features)
}
\arguments{
\item{features}{sf features}
}
\description{
Update map for polygon mode
}
