% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requestPop.R
\name{requestPop}
\alias{requestPop}
\title{Get population estimate from GRID3 server via API request}
\usage{
requestPop(geojson, iso3, ver, timeout = 30, test = F)
}
\arguments{
\item{geojson}{A geoJSON to represent the polygon where a population estimate is needed}

\item{iso3}{The ISO3 country code}

\item{ver}{The version of the population estimate}

\item{timeout}{Seconds until timeout}

\item{test}{Logical indicating whether to use the test server or production server}
}
\value{
A vector of samples from posterior distribution of the population total. If the task times out the function will return the task ID.
}
\description{
Get population estimate from GRID3 server via API request
}
