% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveResults.R
\name{retrieveResults}
\alias{retrieveResults}
\title{Retrieve results from WOPR}
\usage{
retrieveResults(
  tasks,
  url_queue,
  confidence = 0.95,
  tails = 2,
  abovethresh = NA,
  belowthresh = NA,
  summarize = T,
  verbose = T,
  save_messages = F
)
}
\arguments{
\item{tasks}{A data frame with information about tasks (see ?submitTasks)}

\item{url_queue}{The url of the WorldPop queue to check for results (see ?endpoint)}

\item{confidence}{The confidence level for the confidence intervals (e.g. 0.95 = 95 percent confidence intervals)}

\item{tails}{The number of tails for the confidence intervals}

\item{abovethresh}{The function will return the probability that the population size exceeds _abovethresh_}

\item{belowthresh}{The function will return the probability that the population size is less than _belowthresh_}

\item{summarize}{Logical indicating to summarize results or return all posterior samples}

\item{verbose}{Logical indicating to print progress updates}

\item{save_messages}{Save messages from WOPR including task ids that can be used to retrieve results later using checkTask()}
}
\value{
A data frame with outputs
}
\description{
Retrieve results from WOPR
}
