% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitTasks.R
\name{submitTasks}
\alias{submitTasks}
\title{Submit features to WorldPop API to process population totals}
\usage{
submitTasks(features, country, ver, agesex, url, key, verbose = T)
}
\arguments{
\item{features}{An object of class sf with polygons or points}

\item{country}{ISO-3 code for the country requested}

\item{ver}{Version number of population estimates}

\item{url}{The url of the WorldPop API endpoint}
}
\value{
A data frame with information about each task submitted
}
\description{
Submit features to WorldPop API to process population totals
}
