% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPop.R
\name{summaryPop}
\alias{summaryPop}
\title{Get population estimate from GRID3 server via API request}
\usage{
summaryPop(N, alpha = 0.05, tails = 2)
}
\arguments{
\item{N}{Vector of posterior samples for the population total}

\item{alpha}{The type 1 error rate for the confidence intervals}

\item{tails}{The number of tails for the confidence intervals}
}
\value{
A list with the mean, median, lower and upper confidence interval for the population total
}
\description{
Get population estimate from GRID3 server via API request
}
