% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPop.R
\name{summaryPop}
\alias{summaryPop}
\title{Summary statistics}
\usage{
summaryPop(N, confidence = 0.95, tails = 2, popthresh = NA)
}
\arguments{
\item{N}{Vector of posterior samples for the population total}

\item{confidence}{The confidence level for the confidence intervals (e.g. 0.95 = 95 percent confidence intervals)}

\item{tails}{The number of tails for the confidence intervals}

\item{popthresh}{}
}
\value{
A list with the mean, median, lower and upper confidence intervals for the population total (rounded to integers)
}
\description{
Summarize predicted posterior probability distribution for population estimates
}
