% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulateTotals.R
\name{tabulateTotals}
\alias{tabulateTotals}
\title{Tabulate population totals for polygons}
\usage{
tabulateTotals(shapes, alpha = 0.05, tails = 2)
}
\arguments{
\item{shapes}{SpatialPolygonsDataFrame with polygons to calculate population totals}
}
\value{
The SpatialPolygonsDataFrame with population totals added into columns: meanPop, medianPop, lowerPop, upperPop
}
\description{
Tabulate population totals for polygons
}
