% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulateTotals.R
\name{tabulateTotals}
\alias{tabulateTotals}
\title{Tabulate population totals for multiple polygons.
Note: The operation may take a number of seconds per polygon.}
\usage{
tabulateTotals(polygons, country, ver, alpha = 0.05, tails = 2,
  timeout = 10 * 60)
}
\arguments{
\item{polygons}{SpatialPolygonsDataFrame with polygons to calculate population totals}

\item{country}{ISO-3 code for the country requested}

\item{ver}{Version number of population estimates}

\item{alpha}{The type 1 error rate for the confidence intervals}

\item{tails}{The number of tails for the confidence intervals}

\item{timeout}{Seconds until the operation for a single polygon times out}
}
\value{
A data frame with summaries of posterior distribtuions for estimates of total population within each polygon
}
\description{
Tabulate population totals for multiple polygons.
Note: The operation may take a number of seconds per polygon.
}
