% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woprize.R
\name{woprize}
\alias{woprize}
\title{WOPRize a shapefile}
\usage{
woprize(
  features,
  country,
  ver = NA,
  confidence = 0.95,
  tails = 2,
  popthresh = NA,
  spatialjoin = T,
  summarize = T,
  timeout = 30 * 60,
  agesex = c("m0", "m1", "m5", "m10", "m15", "m20", "m25", "m30", "m35", "m40", "m45",
    "m50", "m55", "m60", "m65", "m70", "m75", "m80", "f0", "f1", "f5", "f10", "f15",
    "f20", "f25", "f30", "f35", "f40", "f45", "f50", "f55", "f60", "f65", "f70", "f75",
    "f80"),
  key = "key.txt"
)
}
\arguments{
\item{features}{An object of class sf with points or polygons to calculate population totals}

\item{country}{ISO-3 code for the country requested}

\item{ver}{Version number of population estimates}

\item{confidence}{The confidence level for the confidence intervals (e.g. 0.95 = 95 percent confidence intervals)}

\item{tails}{The number of tails for the confidence intervals}

\item{popthresh}{Threshold population size to calculate the probability that population exceeds}

\item{summarize}{Logical indicating to summarize results or return all posterior samples}

\item{timeout}{Seconds until the operation for a single polygon times out}

\item{agesex}{Character vector of age-sex groups}

\item{key}{API key}

\item{spatialJoin}{Logical indicating to join results to sf spatial data or to return a data frame}
}
\value{
A data frame or sf spatial data object with summaries of posterior distribtuions for estimates of total population within each polygon
}
\description{
Query WOPR to get population totals and confidence intervals for sf polygon or point features.
}
