% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woprize.R
\name{woprize}
\alias{woprize}
\title{WOPRize a shapefile}
\usage{
woprize(
  features,
  country,
  version = NA,
  confidence = 0.95,
  tails = 2,
  abovethresh = NA,
  belowthresh = NA,
  summarize = T,
  agesex_select = c(paste0("m", c(0, 1, seq(5, 80, 5))), paste0("f", c(0, 1, seq(5, 80,
    5)))),
  key = "key.txt",
  save_messages = F,
  url = "https://api.worldpop.org/v1"
)
}
\arguments{
\item{features}{An object of class sf with points or polygons to calculate population totals}

\item{country}{ISO-3 code for the country requested}

\item{version}{Version number of population estimates}

\item{confidence}{The confidence level for the confidence intervals (e.g. 0.95 = 95 percent confidence intervals)}

\item{tails}{The number of tails for the confidence intervals}

\item{abovethresh}{The function will return the probability that the population size exceeds _abovethresh_}

\item{belowthresh}{The function will return the probability that the population size is less than _belowthresh_}

\item{summarize}{Logical indicating to summarize results or return all posterior samples}

\item{agesex_select}{Character vector of age-sex groups}

\item{key}{API key}

\item{save_messages}{Save messages from WOPR including task ids that can be used to retrieve results later using checkTask()}

\item{url}{Server url (optional)}
}
\value{
A data frame or sf spatial data object with summaries of posterior distribtuions for estimates of total population within each polygon
}
\description{
Query WOPR to get population totals and confidence intervals for sf polygon or point features.
}
