% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woprVision.R
\name{woprVision}
\alias{woprVision}
\title{Run the woprVision application}
\usage{
woprVision(wopr_dir = "wopr")
}
\arguments{
\item{wopr_dir}{Path to WOPR files stored locally.}
}
\value{
A shiny application will be executed.
}
\description{
Run the woprVision application to visualize results from the WorldPop Open Population Repository.
}
