% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCatalogue.R
\name{writeCatalogue}
\alias{writeCatalogue}
\title{Creates a catalogue of the WOPR data currently in the output directory.}
\usage{
writeCatalogue(outdir)
}
\arguments{
\item{outdir}{Output folder where data from the WorldPop GridFree catalogue are stored}
}
\value{
Writes an updated data catalogue to disk as a .csv file.
}
\description{
Creates a catalogue of the WOPR data currently in the output directory.
}
